Fixed test examples s.t they work with the new performance improvements

+++ src/NewFrontend/NFBinding.jl
@@ -29,7 +29,7 @@
 *
 =#
 
-const VariabilityType = Int
+const VariabilityType = Int8
 
 @UniontypeDecl Binding
